api_yamdb/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1

pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3

py~=1.11.0
colorama~=0.4.4
pip~=21.1.2
wheel~=0.36.2
toml~=0.10.2
attrs~=21.4.0
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
certifi~=2021.10.8
idna~=3.3
urllib3~=1.26.8
pyparsing~=3.0.6
setuptools~=57.0.0